+++ compiler-core/src/warning.rs
@@ -1276,6 +1276,21 @@ can already tell whether it will be true or false.",
                         extra_labels: Vec::new(),
                     }),
                 },
+                type_::Warning::TopLevelDefinitionShadowsImport { location, name } => Diagnostic {
+                    title: format!("Definition of \"{name}\" shadows an imported value"),
+                    text: wrap("The imported value could not be used in functions anyway."),
+                    hint: None,
+                    level: diagnostic::Level::Warning,
+                    location: Some(Location {
+                        src: src.clone(),
+                        path: path.to_path_buf(),
+                        label: diagnostic::Label {
+                            text: None,
+                            span: *location,
+                        },
+                        extra_labels: vec![],
+                    }),
+                },
             },
 
             Warning::DeprecatedEnvironmentVariable { variable } => {